# Append to lists and strings in place instead of cloning when possible