As long as lists and strings are used behind an Arc, the left-hand list can be modified in place when no other reference to it exists.

Additionally, when the same condition applies to the right-hand list, its items can be moved into the left-hand list, rather than cloned.

Together, this can make expressions like `[...] + [...] + [...]` more efficient.

According to `cargo bench` locally:

```
add_list        time:   [180.38 ns 180.58 ns 180.84 ns]
                change: [-16.091% -15.608% -15.162%] (p = 0.00 < 0.05)
                Performance has improved.

add_string      time:   [75.361 ns 75.448 ns 75.535 ns]
                change: [-13.354% -12.974% -12.661%] (p = 0.00 < 0.05)
                Performance has improved.
```

Of course, in both of these benches, both operands are uniquely referenced, and there is only a single thread. That said, from Arc's own documentation (https://doc.rust-lang.org/std/sync/struct.Arc.html):

> Use clone-on-write semantics with Arc::make_mut which provides
efficient mutation without requiring interior mutability. This approach clones the data only when needed (when there are multiple references) and can be more efficient when mutations are infrequent.

This change cannot cause values contained in a Context to be modified by construction: The Context itself retains a strong reference, meaning that the values received by `Add::add` cannot be unique.