@apasel422 2025-06-26
With this longer benchmark, the difference is even more pronounced:

```cel
['a'] + ['b', 'c'] + ['d', 'e', 'f'] + ['g', 'h', 'i', 'j']
```

Before:
```
add_list_strings2       time:   [851.25 ns 860.33 ns 872.00 ns]
```

After:
```
add_list_strings2       time:   [576.51 ns 579.26 ns 582.33 ns]
                        change: [-32.812% -32.155% -31.618%] (p = 0.00 < 0.05)
                        Performance has improved.
```